==> 01_downloads_repositories.ipynb <==
# Repositories Download

Here is processed the download of repositories.

import csv
import pandas as pd
import numpy as np
import os
import git
from git import Repo
from pydriller import GitRepository

### Declaration of constants and paths

PATH_REPOSITORIES = 'repositories'
PATH_RESULTS = 'results'
FILE_LIST_REPOSITORIES = 'data/filtered_data.csv'

### Selects 5 download repositories

df = pd.read_csv(FILE_LIST_REPOSITORIES)
list_repositories = df[:100]
list_repositories

### Create paths

#created path repositories 
if not os.path.exists(PATH_REPOSITORIES):
    os.makedirs(PATH_REPOSITORIES)

#created path results 
if not os.path.exists(PATH_RESULTS):
    os.makedirs(PATH_RESULTS)


for repository in list_repositories.sourceRep :
#     print('cloning repostitory %s...' %(repository))
    
    rep_name = repository.split('https://github.com/',1)[1]
    name_path = rep_name.replace('/','_')
    path_rep_clone = PATH_REPOSITORIES + '/' + name_path
    
    if not (os.path.exists(path_rep_clone)): #clona se não existe
        try:
            print('cloning repostitory %s...' %(repository))
            Repo.clone_from(repository,path_rep_clone)    
        except:
            print('not cloning or repostitory exists %s...' %(repository))
    else:
        print('repository %s exist' %(repository))
print('clone completed...')


#verificar quais repositórios fez download
dataDownload = []
for repository in list_repositories.sourceRep :
    rep_name = repository.split('https://github.com/',1)[1]
    name_path = rep_name.replace('/','_')
    path_rep_clone = PATH_REPOSITORIES + '/' + name_path
    
    
    if (os.path.exists(path_rep_clone)):
        gr = GitRepository(path_rep_clone)
        dataDownload.append({'name' : rep_name, 'link' : repository, 'path' : path_rep_clone, 'commits' : gr.total_commits()})

df = pd.DataFrame(dataDownload)

df

df.to_csv('data/rep_down.csv')
