==> 02_capturing_issues.ipynb <==
# Capituring Repositories Issues

import pandas as pd
import numpy as np
import os
import csv

from github import Github

PATH_REPOSITORIES = 'repositories'
PATH_RESULTS = 'results'
FILE_LIST_REPOSITORIES = 'data/filtered_data.csv'
FILE_DOWN = 'data/rep_down.csv'

df = pd.read_csv(FILE_LIST_REPOSITORIES)
repositories_down = df[:100]

rep_info = []

for rep in repositories_down.values:
    name = rep[1].split('https://github.com/')[1]
    rep_info.append([name,rep[1],PATH_REPOSITORIES + '/' + name.replace('/','_')])
    
dataset = pd.DataFrame(rep_info,columns=['name','link','path'])
dataset



g = Github('matheusflauzino','lola@2014')

for index, row in dataset.iterrows():
    path_result = 'issues/' + row['path'].split('/')[1] + '.csv'
    print(row['link'])
    print(path_result)
    if not (os.path.exists(path_result)):
        repository = g.get_repo(row['name'])
        name = row['name']
#         count_stars = repository.stargazers_count
        
        data_result = [] #clear variable data
         
        try:
            issues = repository.get_issues(state='all') #all issues
            for issue in issues: #interate issues
                labels = getattr(issue,'labels','')
                is_bug = 0
                for label in labels:
                    if label.name == 'bug':
                        is_bug = 1
                data_result.append([name,getattr(issue,'state',''),getattr(issue,'closed_at',''),getattr(issue,'title',''),list(getattr(issue,'labels','')),is_bug,getattr(issue,'body','')])

            if not (os.path.exists(path_result)):
                with open(path_result, 'w') as csvfile:
                    spamwriter = csv.writer(csvfile, delimiter=',', quoting=csv.QUOTE_MINIMAL)
                    spamwriter.writerow(['repository','state','closed_at','title','label','is_bug','body'])
                    spamwriter.writerows(data_result)
        except:
            print('no exist issues')

df = pd.read_csv('issues/gsantner_memetastic.csv')
df

df['state'].value_counts()

downloads = pd.read_csv(FILE_DOWN)

data_issues = []
for repository in downloads.name:
    file_issues = 'issues/' + repository.replace('/','_') + '.csv'
    print(file_issues)
    df_issues = pd.read_csv(file_issues)
    
    #coleta as informações
    values_counts = df_issues['state'].value_counts()
    issues_bugs = 0
    for index, row in df_issues.iterrows() :
        if row['state'] == 'open' and row['is_bug']:
#             print(index , ' ' , row['title'], ' ', row['is_bug'], ' ', getattr(values_counts,'open',0))
            issues_bugs += 1
    issues_no_bugs = getattr(values_counts,'open',0)-issues_bugs
#     print('{}: Open {} - Bugs {} - no-Bugs {}'.format(repository,values_counts.open,issues_bugs,issues_no_bugs))
    data_issues.append({'name':repository, 'open': getattr(values_counts,'open',0), 'bugs' : issues_bugs, 'no_bugs': issues_no_bugs})
        

dataset = pd.DataFrame(data_issues,index=None,columns=['name','open','bugs','no_bugs'])
dataset

dataset.to_csv('data/issues.csv')
