==> 04_pmd_repositories.ipynb <==
# Code Analyzer with PMD

import csv
import os
import pandas as pd
import numpy as np

PATH_REPOSITORIES = 'repositories'
PATH_RESULTS = 'results'
FILE_LIST_REPOSITORIES = 'data/filtered_data.csv'
PMD = '/home/matheus/pmd-bin-6.9.0/bin/run.sh pmd'

#created path repositories 
if not os.path.exists(PATH_REPOSITORIES):
    os.makedirs(PATH_REPOSITORIES)

#created path results 
if not os.path.exists(PATH_RESULTS):
    os.makedirs(PATH_RESULTS)

#load csv repositories data
list_repositories = pd.read_csv(FILE_LIST_REPOSITORIES)
repositories = list_repositories[:5]
repositories

for repository in repositories.sourceRep:
    rep_name = repository.split('https://github.com/',1)[1]
    name_path = rep_name.replace('/','_')
    path_rep_clone = PATH_REPOSITORIES + '/' + name_path
    path_name_repository = os.getcwd() + '/' + path_rep_clone
    path_name_result = os.getcwd() + '/' + PATH_RESULTS + '/' + name_path + '.csv'
    str_command = PMD + ' -d {} -f csv -R "category/java/design.xml/ExcessiveClassLength, category/java/design.xml/ExcessiveMethodLength, category/java/design.xml/ExcessiveParameterList, category/java/design.xml/DataClass, category/java/design.xml/TooManyMethods" -r {}'.format(path_name_repository,path_name_result)
    os.system(str_command)
    print('runing PMD in repository {}...'.format(rep_name))
print('end of execution...')

df = pd.read_csv('results/hidroh_materialistic.csv')

df['Rule'].value_counts()
