==> 05_locs_informations.ipynb <==
import csv
import os
import pandas as pd
import numpy as np

PATH_REPOSITORIES = 'repositories'
PATH_LOC = 'loc_results'
PATH_RESULTS = 'results'
FILE_LIST_REPOSITORIES = 'data/filtered_data.csv'

#load csv repositories data
list_repositories = pd.read_csv(FILE_LIST_REPOSITORIES)
repositories = list_repositories[:100]
repositories

data_loc = []
for repository in repositories.sourceRep:
    rep_name = repository.split('https://github.com/',1)[1]
    name_path = rep_name.replace('/','_')
    path_rep_clone = PATH_REPOSITORIES + '/' + name_path
    path_name_repository = os.getcwd() + '/' + path_rep_clone
    path_result_loc = os.getcwd() + '/' + PATH_LOC + '/' + name_path + '.csv'   
    command_loc = 'cloc --csv {} --out="{}"'.format(path_name_repository,path_result_loc)
    os.system(command_loc)
     #create file brank
    if not (os.path.exists(path_result_loc)):
        with open(path_result_loc, 'w') as csvfile:
            spamwriter = csv.writer(csvfile, delimiter=',', quoting=csv.QUOTE_MINIMAL)
            spamwriter.writerow(['files','language','blank','comment','code'])
            spamwriter.writerows([0,0,0,0,0])
            
    loc_data = pd.read_csv(path_result_loc)
    for index, row in loc_data.iterrows():
        if row['language'] == 'Java':
            data_loc.append({'name' : rep_name, 'loc':row['code']})
#             data_loc[row['index']] = row['code']
            
    

df = pd.DataFrame(data_loc,index=None,columns = ['name','loc'])
df

df.to_csv('data/loc_repositories_total.csv')
