==> 06_result_data.ipynb <==
import csv
import os
import pandas as pd
import numpy as np

PATH_REPOSITORIES = 'repositories'
PATH_LOC = 'loc_results'
PATH_RESULTS = 'results'
FILE_LIST_REPOSITORIES = 'data/filtered_data.csv'
FILE_LOC = 'data/loc_repositories_total.csv'
FILE_DOWN = 'data/rep_down.csv'
FILE_SMELLS = 'data/smells.csv'
FILE_ISSUES = 'data/issues.csv'

downloads = pd.read_csv(FILE_DOWN)

downloads

locs = pd.read_csv(FILE_LOC)

locs

smells = pd.read_csv(FILE_SMELLS)

smells.loc[:,'tmm':'smells']


smells

issues = pd.read_csv(FILE_ISSUES)

issues

df = smells.loc[:,'name':'smells']
df['loc'] = pd.Series(locs['loc'],index=df.index)

df['commits']  = pd.Series(downloads['commits'],index=df.index)
df['link']  = pd.Series(downloads['link'],index=df.index)

df['issues'] = pd.Series(issues['open'],index=df.index)
df['bugs'] = pd.Series(issues['bugs'],index=df.index)
df['no_bugs'] = pd.Series(issues['no_bugs'],index=df.index)

dataset = pd.DataFrame(df,index=None,columns=['name','link','loc','dc','lc','lm','lpl','tmm','smells','issues','bugs','no_bugs'])

dataset

dataset.to_csv('data/dataset_finaly.csv')

dataset['smells'].min()
